+++ app/(main)/(routes)/meself/layout.tsx
@@ -37,6 +37,7 @@ const MePageLayout = async ({
                             src={profile.imageUrl}
                             name={profile.name}
                             status={profile.status}
+                            about={profile.about}
                         />
                        
                     </div>