+++ components/providers/prompt-provider.tsx
@@ -14,6 +14,7 @@ import { EditChannelPrompt } from "../prompts/channel/edit-channel-prompt";
 import { MessageFilePrompt } from "../prompts/file/message-file-modal";
 import { DeleteMessageModal } from "../prompts/message/delete-message-prompt";
 import { UserProfilePrompt } from "../prompts/user/user-profile-prompt";
+import { EditUserPrompt } from "../prompts/user/edit-user-prompt";
 
 export const PromptProvider = () => {
     const [isMounted,setIsMounted] = useState(false);
@@ -42,6 +43,7 @@ export const PromptProvider = () => {
             <DeleteMessageModal />
             <DeleteDirectChannelPrompt />
             <UserProfilePrompt />
+            <EditUserPrompt />
         </>
     )
 }